Abl2 {pendensity}	R Documentation

Calculating the second order derivative with and without penalty

Description

Calculating the second order derivative of the likelihood function of the pendensity approach w.r.t. the parameter beta. Thereby, for later use, the programm returns the second order derivative with and without the penalty.

Usage

Abl2(penden.env, lambda0)

Arguments

penden.env 	 Containing all information, environment of pendensity()

lambda0 	 smoothing parameter lambda

Details

We approximate the second order derivative in ths approch with the negative fisher information.

J(beta)= partial^2 l(beta) / (partial(beta) partial(beta)) = sum(s[i](beta) s[i]^T(beta))

Therefore we construct the second order derivative of the i-th observation w.r.t. beta with the outer product of the matrix Abl1.cal and the i-th row of the matrix Abl1.cal.
The penalty is computed as

lambda Dm

.

Value

Abl2.pen 	 second order derivative w.r.t. beta with penalty

Abl2.cal 	 second order derivative w.r.t. beta without penalty. Needed for calculating of e.g. AIC.

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>

References

Penalized Density Estimation, Kauermann G. and Schellhase C. (2009), to appear.

[Package pendensity version 0.1 Index]