bias.par {pendensity}	R Documentation

Calculating the bias of the paramater beta

Description

Calculating the bias of the parameter beta.

Usage

bias.par(penden.env)

Arguments

penden.env 	 Containing all information, environment of pendensity()

Details

The bias of the parameter beta is calculated as the product of the penalty parameter lambda, the penalzied second order derivative of the log likelihood function w.r.t. beta 'Abl2.pen', the penalty matrix 'Dm' and the parameter set 'beta'.

Bias(beta)=lambda Abl2.pen(beta)^-1 Dm beta

The needed values are saved in the environment.

Value

Returning the bias of the parameter beta.

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>

References

Penalized Density Estimation, Kauermann G. and Schellhase C. (2009), to appear.

[Package pendensity version 0.1 Index]