ck {pendensity}	R Documentation

Calculating the actual weights ck

Description

Calculating the actual weights ck for each factor combination of the covariates combinations.

Usage

ck(penden.env, beta.val)

Arguments

penden.env 	 Containing all information, environment of pendensity()

beta.val 	 actual parameter beta

Details

The weights in depending of the covariate 'x' are calculated as follows.

c_k(x,β)=frac{exp(Z(x)β_k)}{sum_{j=-K}^{K}exp(Z(x)β_j)}

For estimations without covariates, Z doesn't appear in calculations.

c_k(β)=frac{exp(β_k)}{sum_{k=-K}^{K}exp(β_k)}

Starting density calculation, the groupings of the covariates are indexed in the main program. The groupings are saved in 'x.factor', the index which response belongs to which group is saved in 'Z.index'. Therefore, one can link to the rows in 'x.factor' to calculate the weights 'ck'.

The needed values are saved in the environment.

Value

Returning the actual weights ck, depending on the actual parameter beta in a matrix with rows.

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>

References

Penalized Density Estimation, Kauermann G. and Schellhase C. (2009), to appear.

[Package pendensity version 0.1 Index]