distr.func {pendensity}	R Documentation

These functions are used for calculating the empirical and theorectical distribution functions.

Description

These functions cooperate with each other for calculating the distribution functions. 'distr.func' is the main program, calling 'distr.func.help',generating an environment with needed values for calculating the distribution of each interval between two neighbouring knots. 'distr.func' returns analytical functions of the distribution of each interval between two neighbouring knots. Therefore the function 'poly.part' is needed to construct these funtions. 'cal.int' evaluates these intergrals, considering if the whole interval should be evaluated or if any discrete value 'yi' is of interest.

Usage

distr.func(yi = NULL, obj, help.env=distr.func.help(obj))
distr.func.help(obj)
cal.int(len.b, q, help.env, knots.val)
poly.part(i,j,knots.val,help.env,q, yi=NULL, poly=FALSE)

Arguments

yi 	 if the distribution at any discrete point is of interest, you can call for it. Default=NULL doesn't consider any discret point

obj 	 a object of class pendensity

help.env 	 object is generated with calling distr.func.help(obj)

len.b 	 length of B-Spline

q 	 order of the B-Spline

knots.val 	 values of the used knots

poly 	 TRUE/FALSE

i 	 internal values for calculating the polynoms of each B-Spline

j 	 internal values for calculating the polynoms of each B-Spline

Value

distr.func 	 returns analytical functions of the distributions between each two neighbouring intervals

distr.func.help 	 creating environment 'help.env', creating help points between each two neighbouring knots and calcutes the polynomial-coefficients of each base part

cal.int 	 evaluting the result of distr.func. Thereby it's possible to call for an explicit distribution values F(yi)

poly.part 	 using in 'distr.func' for creating the polynomial functions of each interval of each two neighbouring knots

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>

References

Penalized Density Estimation, Kauermann G. and Schellhase C. (2009), to appear.

[Package pendensity version 0.1 Index]