f.hat {pendensity}	R Documentation

Calculating the actual fitted values of y

Description

Calculating the actual fitted values of the response, depending on the actual parameter set beta

Usage

f.hat(penden.env,ck.temp=NULL)

Arguments

penden.env 	 Containing all information, environment of pendensity()

ck.temp 	 actual weights, depending on the actual parameter set beta. If NULL, the beta parameter is caught in th environment

Details

Calculating the actual fitted values of the response, depending on the actual parameter set beta. Multiplying the actual set of parameters c_k with the base 'base.den' delivers the fitted values, depending on the group of covariates, listed in 'x.factor'.

Value

The returned value is a vector of the fitted value for each observation of y.

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>

References

Penalized Density Estimation, Kauermann G. and Schellhase C. (2009), to appear.

[Package pendensity version 0.1 Index]