my.AIC {pendensity}	R Documentation

Calculating the AIC value

Description

Calculating the AIC vaule of the density estimation. Therefore, we add the unpenalized log likelihood of the estimation and the degree of freedom, which are

Usage

my.AIC(penden.env, lambda0, opt.Likelihood = NULL)

Arguments

penden.env 	 Containing all information, environment of pendensity()

lambda0 	 penalty parameter lambda

opt.Likelihood 	 optimal unpenalized likelihood of the density estimation

Details

AIC is calculated as AIC(λ)= - l(hat{β}) + df(λ)

Value

myAIC 	 sum of the negative unpenalized log likelihood and mytrace

mytrace 	 calculated mytrace as the sum of the diagonal matrix df, which results as the product of the invers of the penalized second order derivative of the log likelihood with the unpenalized second order derivative of the log likelihood

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>

References

Penalized Density Estimation, Kauermann G. and Schellhase C. (2009), to appear.

[Package pendensity version 0.1 Index]