my.positive.definite.solve {pendensity}	R Documentation

my.positive.definite.solve

Description

Invers a quadratic positive definite matrix.

Usage

my.positive.definite.solve(A, eps = 1e-15)

Arguments

A 	 quadratic positive definite matrix

eps 	 level of the lowest eigenvalue to consider

Details

The programm makes an eigenvalue decomposition of the positive definite matrix A and searchs all eigenvalues greater than eps. The value of return is the invers matrix of A, constructed with the matrix product of the corresponding eigenvalues and eigenvectors.

Value

The return is the invers matrix of A.

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>

[Package pendensity version 0.1 Index]