new.beta.val {pendensity}	R Documentation

Calculating the new parameter beta

Description

Calculating the direction of the Newton-Raphson step for the known beta and iterate a step size bisection to control the maximizing of the penalized likelihood.

Usage

new.beta.val(llold, penden.env)

Arguments

llold 	 log likelihood of the algorithm one step before

penden.env 	 Containing all information, environment of pendensity()

Details

We terminate the search for the new beta, when the new log likelihood is smaller than the old likelihood and the step size is smaller or equal 1e-3. We calculate the direction of the Newton Raphson step for the known beta_t and iterate a step size bisection to control the maximizing of the penalized likelihood

l(beta,lambda0)

. This means we set

beta[t+1]=beta[t]-(2/v)*sp(beta,lambda0)*(-Jp(beta[t],lambda0))^-1

with s_p as penalized first order derivative and J_p as penalized second order derivative. We begin with v=0. Not yielding a new maximum for a current v, we increase v step by step respectively bisect the step size. We terminate the iteration, if the step size is smaller than some reference value epsilon (eps=1e-3) without yielding a new maximum. We iterate for new parameter beta until the new log likelihood depending on the new estimated parameter beta differ less than 0.1 log-likelihood points from the log likelihood estimated before.

Value

Likelie 	 corresponding log likelihood

step 	 used step size

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>

References

Penalized Density Estimation, Kauermann G. and Schellhase C. (2009), to appear.

[Package pendensity version 0.1 Index]