new.lambda {pendensity}	R Documentation

Calculating new penalty paramater lambda

Description

Calculating new penalty parameter lambda.

Usage

new.lambda(penden.env,lambda0)

Arguments

penden.env 	 Containing all information, environment of pendensity()

lambda0 	 actual penalty parameter lambda

Details

Iterating for the lambda is stopped, when the changes between the old and the new lambda is smaller than 0.01*lambda0. If this criterion isn't reached, the iteration is terminated after 11 iterations.

The iteration formulae is

λ=(β^T D_m β)/(df(λ)-(m-1)

Value

Returning the new lambda.

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>

References

Penalized Density Estimation, Kauermann G. and Schellhase C. (2009), to appear.

[Package pendensity version 0.1 Index]