pen.log.like {pendensity}	R Documentation

Calculating the log likelihood

Description

Calculating the considered log likelihood. If one chooses lambda0=0, one gets the (actual) unpenalized log likelihood. Otherwise, one gets the penalized log likelihood for the used fitted values of the response y and the actual parameter set beta.

Usage

pen.log.like(penden.env,lambda0,f.hat.val=NULL,beta.val=NULL)

Arguments

penden.env 	 Containing all information, environment of pendensity()

lambda0 	 penalty parameter lambda

f.hat.val 	 matrix contains the fitted values of the response, if NULL the matrix is caught in the environment

beta.val 	 actual parameter set beta, if NULL the vector is caught in the environment

Details

The calculation depends on the fitted values of the response as well as on the penalty parameter lambda and the penalty matrix Dm.

l(β)=sum(log(sum(c_k(x_i,β) phi_k(y_i))))-0.5*λ β^T D_m β

.

The needed values are saved in the environment.

Value

Returns the log likelihood depending on the penalty parameter lambda.

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>

References

Penalized Density Estimation, Kauermann G. and Schellhase C. (2009), to appear.

[Package pendensity version 0.1 Index]