pendenForm {pendensity}	R Documentation

Formula interpretation and data transfer

Description

Function 'pendenForm' interprets the input 'form' of the function pendensity(),transfers the response and covariates data back to the main program and categorize the values to groupings.

Usage

pendenForm(form)
string.help(string, star = " ")

Arguments

form 	 formula from the input of pendensity()

string 	 string of the formula

star 	 seperating letter

Value

Returing the values and the structure of response and covariates.

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>

References

Penalized Density Estimation, Kauermann G. and Schellhase C. (2009), to appear.

[Package pendensity version 0.1 Index]