test.equal {pendensity}	R Documentation

Testing pairwise equality of densities

Description

Every group of factor combination is tested pairwise for equality to all other groups.

Usage

test.equal(obj)

Arguments

obj 	 object of class pendensity

Details

We consider the distribution of the integrated B-Spline density base. This is saved in the program in the object named 'mat1'. Moreover, we use the variance 'var.par' of the estimation, the weights and some matrices 'C' of the two compared densities to construct the matrix 'W'. We simulate the distribution of the test statistic using a spectral decomposition of W.

Value

Returning a list of p-values for testing pairwise for equality.

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld>

References

Penalized Density Estimation, Kauermann G. and Schellhase C. (2009), to appear.

[Package pendensity version 0.1 Index]