variance.par {pendensity}	R Documentation

Calculating the variance of the parameters

Description

Calculating the variance of the parameters of the estimation, depending on the second order derivative and the penalized second order derivative of the density estimation.

Usage

variance.par(penden.env)

Arguments

penden.env 	 Containing all information, environment of pendensity()

Details

The variance of the parameters of the estimation results as the product of the invers of the penalized second order derivative times the second order derivative without penalization time the invers of the penalized second order derivative.

V(β, λ_0)=I_p^{-1}(β, λ) I_p(β, λ=0) I_p^{-1}(β, λ) with I_p(β^{-1}, λ)=E_{f(y)}bigl{J_p(β, λ)bigr}

The needed values are saved in the environment.

Value

The return is a variance matrix of the dimension (K-1)x(K-1).

Author(s)

Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>

References

Penalized Density Estimation, Kauermann G. and Schellhase C. (2009), to appear.

[Package pendensity version 0.1 Index]