variance.val {pendensity}	R Documentation

Calculating the variances and standard deviance of each observation.

Description

Calculating the variances and standard deviance of each observation for which 'base.den' has been created. Therefore we use the variance of the parameter set beta 'var.par'.

Usage

variance.val(base.den, var.par, weight, K, x, list.len, Z, x.factor)

Arguments

base.den 	 base values

var.par 	 variance of the parameter set beta

weight 	 weights ck

K 	 number of knots

x 	 covariates

list.len 	 number of covariate combinations

Z 	 covariate matrix

x.factor 	 list of covariate combinations

Value

Returning a vector with the standard deviance of each observation for which 'base.den' has been created.

[Package pendensity version 0.1 Index]